// src/utils/Interfaces/Notes.ts
export interface Notes {
    id?: string;
    created_at?: string;
    user_id?: string;
    procedure?: string;
    date_procedure?: Date;
    patient_name?: string;
    doctor_name?: string;
    size_anesthetic?: string;
    laser?: string;
    number_shots?: number;
    application_time?: number;
    place_application?: string;
    comments?: string;
    time_procedure?: string;
}